# examples/sample_openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sample Marketplace API",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "hunter2"
          }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate a user",
        "security": [],
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
...
  }
}